refactor(slacksvc): move run-state DM seam to notifier_state.go (PRD #1026 M2)
Pure same-package motion: move the run-state DM seam out of notifier.go
into the new notifier_state.go, in original source order. Moved verbatim:

State handlers: handle, handleQuestion, handleGate, handleMilestone.
Renderers/helpers: rootBlocks, mrContextElem, decodeMilestones,
decodeMilestoneIDs, milestoneCounts, inProgressTitle, renderThreadBlocks,
cancelledThreadBlocks, limitWaitDetail, statusGlyph, mrLink, isHTTPSURL,
forgeMrAbbrev, forgeMrRef, iid.

Positional-comment fix in handle's body: reworded "after the
render/handleChat below on all paths" to "after the render/handleChat
call on all paths" (handleChat moved to notifier_chat.go in M1, so
"below" was false). Body comment only; go doc output is unaffected.

Left in notifier.go: logf, threadSectionBlock, threadMrkdwnElem,
maxFailureReason, boundReason, runURL, runLink, plus the queue/struct/
event/Publish*/Run scaffold from M1's stays.

go doc -all and go doc -u -all stay byte-identical to baseline.
